Support unclassified folders in ON instrument output
The Oxford Nanopore instrument output provides unclassified
folders that contain read information with reads that had either
no or an unknown sample code to the experiment.

This commit updates the data-model-lib dependency to 1.9.0, which
already supports the unclassified folder structure.

Also, the schema JSON has been removed from the resources and the
Schema is now accessed via the data-model-lib, as the schema will be
maintained there in the future.